Move DecimalIter to iterator.rs
